Fix playback queue of audio for analysis in strict browsers
Some browsers such as Chrome 62 on Android 7.1, and Safari on iOS 11
prohibits to queue playback with play event.
To play audio for analysis on such environments, its playback must be
queued in the context audio for output gets queued, which is guranteed to
be capable to queue playback. Pausing operations except one which
automatically happens after the stream end are removed for the
restriction.